Removed blacklist token logic, added UUID to User model, updated auth handlers
This commit removes the previously implemented blacklisted token logic and the accompanying model altogether. Instead, the responsibility of deleting JWTs in the frontend is given to the client during logging out, improving efficiency. A UUID field was also added to the User model which now acts as the primary key. JWT generation for new users was added during signup and an external JWT secret key is being used in both login and signup.