Refactor database and dotenv configuration
The sequelizeOptions for database has been refactored to be simpler and more readable. SSL usage is now determined from a configuration file instead of detecting if the server is in production mode. Also adjusted dotenv configuration to use the default path instead of hardcoding the .env path. Additionally, scripts to synchronize the database in production and development environments are added in package.json.